+++ packages/jaspr_builder/lib/src/client/client_registry_builder.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:build/build.dart';
-import 'package:dart_style/dart_style.dart';
 import 'package:glob/glob.dart';
 
 import '../utils.dart';
@@ -18,7 +17,7 @@ class ClientRegistryBuilder implements Builder {
       var source = await generateClients(buildStep);
       if (source != null) {
         var outputId = AssetId(buildStep.inputId.package, 'web/main.clients.dart');
-        await buildStep.writeAsString(outputId, source);
+        await buildStep.writeAsFormattedDart(outputId, source);
       }
     } catch (e, st) {
       print('An unexpected error occurred.\n'
@@ -35,9 +34,6 @@ class ClientRegistryBuilder implements Builder {
         'lib/\$lib\$': ['web/main.clients.dart']
       };
 
-  String get generationHeader => "// GENERATED FILE, DO NOT MODIFY\n"
-      "// Generated with jaspr_builder\n";
-
   Future<String?> generateClients(BuildStep buildStep) async {
     var modules = buildStep
         .findAssets(Glob('lib/**.client.json'))
@@ -53,8 +49,6 @@ class ClientRegistryBuilder implements Builder {
     }
 
     var source = '''
-      $generationHeader
-      
       import 'package:jaspr/browser.dart';
       [[/]]
       
@@ -68,10 +62,6 @@ class ClientRegistryBuilder implements Builder {
     ''';
 
     source = ImportsWriter(deferred: true).resolve(source);
-    source = DartFormatter(
-      languageVersion: DartFormatter.latestShortStyleLanguageVersion,
-      pageWidth: 120,
-    ).format(source);
     return source;
   }
 }