+++ packages/jaspr_builder/test/sync/sources/sync_multi.dart
@@ -34,7 +34,10 @@ final syncMultiSources = {
 };
 
 final syncMultiOutputs = {
-  'site|lib/component.sync.dart': '// GENERATED FILE, DO NOT MODIFY\n'
+  'site|lib/component.sync.dart': '// dart format off\n'
+      '// ignore_for_file: type=lint\n'
+      '\n'
+      '// GENERATED FILE, DO NOT MODIFY\n'
       '// Generated with jaspr_builder\n'
       '\n'
       'import \'package:jaspr/jaspr.dart\';\n'
@@ -56,10 +59,7 @@ final syncMultiOutputs = {
       '\n'
       '  @override\n'
       '  Map<String, dynamic> getState() {\n'
-      '    return {\n'
-      '      \'a\': a,\n'
-      '      \'b\': b,\n'
-      '    };\n'
+      '    return {\'a\': a, \'b\': b};\n'
       '  }\n'
       '\n'
       '  @override\n'
@@ -85,10 +85,7 @@ final syncMultiOutputs = {
       '\n'
       '  @override\n'
       '  Map<String, dynamic> getState() {\n'
-      '    return {\n'
-      '      \'c\': c,\n'
-      '      \'d\': d,\n'
-      '    };\n'
+      '    return {\'c\': c, \'d\': d};\n'
       '  }\n'
       '\n'
       '  @override\n'
@@ -96,5 +93,6 @@ final syncMultiOutputs = {
       '    super.initState();\n'
       '    SyncStateMixin.initSyncState(this);\n'
       '  }\n'
-      '}\n',
+      '}\n'
+      '',
 };